+++ src/routes/+layout.ts
@@ -1,4 +1,4 @@
-import { dev } from '$app/environment';
+import { dev } from '$app/env';
 import { injectAnalytics } from '@vercel/analytics/sveltekit';
 
 injectAnalytics({ mode: dev ? 'development' : 'production' });